{"types":"C","runs":23}

+++ problem22.java
@@ -3,8 +3,8 @@ import acm.program.*;
 
 public class problem22 extends GraphicsProgram{
 	public void run(){
-		double y = 0;
-		double x = 0;
+		int y = 0;
+		int x = 0;
 		double sizeOfSquare = getHeight()/SIZE_OF_TABLE;
 		GRect square = new GRect(10, 10);
 		for(int i = 0; i < 8; i++){
